+++ backend/src/main/java/io/metersphere/performance/engine/producer/LoadTestProducer.java
@@ -4,10 +4,12 @@ import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
 
 @Service
+@Transactional(rollbackFor = Exception.class)
 public class LoadTestProducer {
     private static final String SEPARATOR = " ";
